requirements.txt
yfinance>=0.2.22
openai>=1.3.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
seaborn>=0.12.0
plotly>=5.15.0
ta>=0.10.2 